requirements.txt
beautifulsoup4==4.8.2
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-registration==2.4.1
django-tinymce==2.8.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
soupsieve==1.9.5
whitenoise==5.0.1
